requirements.txt
pillow>=9.3.0
pystray>=0.19.4
python-dotenv>=1.0.0
requests>=2.28.1
schedule>=1.2.0
screen_brightness_control>=0.20.0